feat: Add program for combining list and dictionary together
This commit adds a new program that combines a list and a dictionary together. The program prompts the user to enter the number of objects to add to the list, and then runs a loop to enter the objects in the list. The program also includes parameters for name, age, and gender, and allows for updating parameters such as profession, location, and education. It provides a menu with options to show and update by name, show and update by index, search for a student, delete a student, add a new student, restart the program, and close the program. In case of multiple users with the same name, the program displays a menu of all users with the same name and allows the user to select which user to update. The program also includes robust error handling and the ability to show all students in the list.

This commit addresses the following user stories:
- As a user, I want to be able to combine a list and a dictionary together.
- As a user, I want to be able to add, remove, view, and update students in the list.
- As a user, I want to be able to search for a student by name.
- As a user, I want to be able to delete a student from the list.
- As a user, I want to be able to add a new student to the list.
- As a user, I want to be able to restart the program.
- As a user, I want to be able to close the program.

This commit also includes the following recent commits:
- chore: Update dictionary values and remove unnecessary code
- Refactor student functions and add global variables
- Update execution count and remove unnecessary code in Python_Class.ipynb
- Add code for removing 'Manav' from the list
- Update execution count and odds list in Python_Class.ipynb
- Add code for generating lists of even and odd numbers without using loops
- Add code to remove all occurrences of number 9 from a list
- ci: add .deepsource.toml
- Add student management functionality
- Revert "Add student management functionality"